requirements.txt
Django==3.2.9
psycopg2==2.9.2
django-environ==0.4.5 
django-redis-cache==3.0.0
aio-pika==6.8